+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/pebble/PebbleSupport.java
@@ -12,6 +12,7 @@ import java.util.Iterator;
 import java.util.UUID;
 
 import nodomain.freeyourgadget.gadgetbridge.GBApplication;
+import nodomain.freeyourgadget.gadgetbridge.R;
 import nodomain.freeyourgadget.gadgetbridge.impl.GBDevice;
 import nodomain.freeyourgadget.gadgetbridge.model.Alarm;
 import nodomain.freeyourgadget.gadgetbridge.model.CalendarEventSpec;
@@ -113,6 +114,16 @@ public class PebbleSupport extends AbstractSerialDeviceSupport {
 
     @Override
     public void onNotification(NotificationSpec notificationSpec) {
+        String currentPrivacyMode = GBApplication.getPrefs().getString("pebble_pref_privacy_mode", getContext().getString(R.string.p_pebble_privacy_mode_off));
+        if (getContext().getString(R.string.p_pebble_privacy_mode_complete).equals(currentPrivacyMode)) {
+            notificationSpec.body = null;
+            notificationSpec.sender = null;
+            notificationSpec.subject = null;
+            notificationSpec.title = null;
+            notificationSpec.phoneNumber = null;
+        } else if (getContext().getString(R.string.p_pebble_privacy_mode_content).equals(currentPrivacyMode)) {
+            notificationSpec.sender = "\n\n\n\n\n" + notificationSpec.sender;
+        }
         if (reconnect()) {
             super.onNotification(notificationSpec);
         }
@@ -120,6 +131,14 @@ public class PebbleSupport extends AbstractSerialDeviceSupport {
 
     @Override
     public void onSetCallState(CallSpec callSpec) {
+        String currentPrivacyMode = GBApplication.getPrefs().getString("pebble_pref_privacy_mode", getContext().getString(R.string.p_pebble_privacy_mode_off));
+        if (getContext().getString(R.string.p_pebble_privacy_mode_complete).equals(currentPrivacyMode)) {
+            callSpec.name = null;
+            callSpec.number = null;
+        } else if (getContext().getString(R.string.p_pebble_privacy_mode_content).equals(currentPrivacyMode)) {
+            callSpec.name = null;
+        }
+
         if (reconnect()) {
             if ((callSpec.command != CallSpec.CALL_OUTGOING) || GBApplication.getPrefs().getBoolean("pebble_enable_outgoing_call", true)) {
                 super.onSetCallState(callSpec);